brandme_core/requirements.txt
# Brand.Me Core - Shared Utilities
# Requirements for brandme_core package

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
httpx==0.25.1
pydantic==2.5.0

# Google Cloud - Spanner
google-cloud-spanner==3.40.1

# Google Cloud - Firestore
google-cloud-firestore==2.14.0

# Observability dependencies
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
